Clean test_alter
Also move table defs to schema_alter module
